requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.9
Markdown==3.6
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
redis==5.1.0
requests==2.32.3
sqlparse==0.5.0
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
